requirements.txt
# ============================================================
#  Skillifly — Python Dependencies
#  Install with: pip install -r requirements.txt
#
#  Database note: currently using SQLite (default).
#  When migrating to PostgreSQL, add: psycopg2-binary==2.9.9
# ============================================================

# Core Django framework
Django==5.2.8

# WSGI server for production (replaces `runserver`)
gunicorn==21.2.0

# Load environment variables from .env file
python-dotenv==1.2.1

# Efficient static file serving (production, without a separate nginx step for /static/)
whitenoise==6.6.0

# Django authentication (email/social login)
django-allauth==65.15.1

# Async task queue (PDF export)
celery==5.6.3
redis==7.4.0
django-celery-results==2.6.0

# HTTP request library (used internally)
requests==2.32.5

# Image processing (user avatars / portfolio images)
Pillow==12.0.0

# HTML parsing utilities
beautifulsoup4==4.14.2
lxml==6.0.2

# CORS headers (API access if needed)
django-cors-headers==4.3.1
